+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Do not force a newline after the doctype in `HTMLRewriter.handle_decl()`.
 * Moved `test_extracted_archive_display` from the unit tests to a separate script.
 * Removed `test_extracted_archive_display`'s dependency on Tkinter.
+### Fixed
+* Rewrite URLs in inline CSS code when extracting.
 
 ## [0.3.2] - 2021-09-26
 ### Added